+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.0.1] - 2020-05-03
+
+### Added
+
+- Support for Yoast 14
+- Breadcrumbs for post type and taxonomies
+- canonical url for post type and taxonomies
+
 ## [3.3.0] - 2020-04-01
 
 ### Added
@@ -14,7 +22,7 @@ Support for Yoast configuration data including:
 - Webmaster verification
 - Social profiles
 - Schemas
-- Breadcrumbs
+- Breadcrumbs (global config)
 
 ## [3.2.0] - 2020-03-09